feat(directory): refactor directory part 2
