+++ package.json
@@ -50,6 +50,7 @@
     "@babel/plugin-proposal-decorators": "7.12.12",
     "@commitlint/cli": "11.0.0",
     "@commitlint/config-conventional": "11.0.0",
+    "@next/bundle-analyzer": "10.0.6",
     "@types/axios": "0.14.0",
     "@types/cookie": "0.4.0",
     "@types/cookie-signature": "1.0.3",
@@ -91,7 +92,8 @@
     "stylelint-prettier": "1.1.2",
     "stylelint-scss": "3.18.0",
     "stylelint-webpack-plugin": "2.1.1",
-    "typescript": "4.1.3"
+    "typescript": "4.1.3",
+    "webpack-visualizer-plugin": "0.1.11"
   },
   "config": {
     "commitizen": {
