+++ src/app/views/movie-detail/app.ts
@@ -1,0 +1,44 @@
+import { GetServerSideProps } from 'next';
+
+import { IMovieDetailProps } from '@/app/views/movie-detail/interface';
+import { VerifiedIsNotEmpty } from '@/library/helper/validator.helper';
+import { IMovieDetail as Detail } from '@/library/model/movie/interface/movie-detail.interface';
+import { csrfToken, setupWeb } from '@/library/modules/csrf';
+import { Get } from '@/library/modules/fetch-api';
+import { IFetchAPIWithoutData } from '@/library/modules/fetch-api/interface/fetch-api.interface';
+import { INextAPIResponse as API } from '@/library/modules/next-api/interface/next-api.interface';
+
+/**
+ * Get Server Side Props Lifecycle
+ * @param {GetServerSidePropsContext} context - context page
+ * @returns {Promise<GetServerSidePropsResult>}
+ */
+export const getServerSideProps: GetServerSideProps = async (param) => {
+  /**
+   * Get Movie Detail Props
+   * @returns {Promise<IMovieDetailProps>}
+   */
+  const GetMovieDetailProps = (movieID: string): Promise<IMovieDetailProps> =>
+    (Get as IFetchAPIWithoutData<API<Detail>>)(
+      `http://localhost:3000/api/movie/detail`,
+      {
+        params: {
+          id: movieID
+        }
+      }
+    )
+      .then(({ data: { code, data } }) => {
+        if (code === 200 && VerifiedIsNotEmpty(data)) {
+          return {
+            movie: data
+          };
+        }
+
+        throw new Error(`Movie not found`);
+      })
+      .catch(() => ({}));
+
+  return setupWeb(async ({ params: { movieID } }) => ({
+    props: { ...(await GetMovieDetailProps(`${movieID}` || ``)), csrfToken }
+  }))(param);
+};
