+++ src/app/views/movie-detail/interface/movie-props.interface.ts
