+++ src/app/views/movie-detail/store/index.ts
