+++ src/library/modules/fetch-api/interface/fetch-api.interface.ts
@@ -1,4 +1,4 @@
-import { AxiosResponse } from 'axios';
+import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 
 /**
  * Fetch API Response
@@ -16,3 +16,26 @@ export interface IFetchAPIResponse<T> {
  * @since 2021.01.25
  */
 export type IFetchAPIPromise<T> = Promise<IFetchAPIResponse<T>>;
+
+/**
+ * Fetch With Data Interface
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @since 2021.02.05
+ */
+export type IFetchAPIWithoutData<Response> = (
+  url: string,
+  config?: AxiosRequestConfig,
+  instance?: AxiosInstance
+) => IFetchAPIPromise<Response>;
+
+/**
+ * Fetch With Data Interface
+ * @author Irfan Andriansyah <irfan@99.co>
+ * @since 2021.02.05
+ */
+export type IFetchAPIWithData<Response, Parameter> = (
+  url: string,
+  data: Parameter,
+  config?: AxiosRequestConfig,
+  instance?: AxiosInstance
+) => IFetchAPIPromise<Response>;
